feat: adding spinner for UI completeness and bug tracking when approval doesn't go through. Also moving validation report to execution under genai/SMA
