+++ pipelines/genai_mapping/databricks.yml
@@ -54,7 +54,7 @@ targets:
   prod:
     mode: production
     variables:
-      pipeline_version: ${var.git_tag}
+      pipeline_version: ${var.git_commit}
       DB_workspace: "staging_sst_01"
       genai_slack_webhook_id: "0fb252f8-f343-4be7-91de-b55f5408a031"  # edvise-data-crew channel on staging_sst_01
     resources:
@@ -63,12 +63,12 @@ targets:
           git_source:
             git_url: https://github.com/datakind/edvise
             git_provider: gitHub
-            git_tag: ${var.git_tag}
+            git_commit: ${var.git_commit}
         edvise_genai_mapping_execute_pipeline:
           git_source:
             git_url: https://github.com/datakind/edvise
             git_provider: gitHub
-            git_tag: ${var.git_tag}
+            git_commit: ${var.git_commit}
 
   # prod:
   #   mode: production
